+++ src/components/Analysis/DetailSheet/index.tsx
@@ -4,8 +4,8 @@ import {
   S_DetailWrapper,
   S_InputCalorie,
   S_SendBtn,
-} from '../../pages/Analysis/style';
-import { S_FlexBox } from '../../style/FlexBox.style';
+} from './style';
+import { S_FlexBox } from '../../../style/FlexBox.style';
 import React, { useState } from 'react';
 
 const DetailSheet = () => {
@@ -21,6 +21,7 @@ const DetailSheet = () => {
 
   const [totalCalorie, setTotalcalorie] = useState(595); // TODO 각 칼로리의 합으로 계산하기
 
+  // 칼로리 정보
   const CalorieDetail = () => {
     return (
       <S_CalorieDetailWrapper>
@@ -41,12 +42,18 @@ const DetailSheet = () => {
 
   return (
     <S_FlexBox flexDirection="column" width="300px" height="500px">
+      {/* 알림 말풍선 */}
       <S_CheckInform>
         <span>칼로리 정보를 확인하고 수정해주세요!</span>
       </S_CheckInform>
+
+      {/* 디테일 */}
       <S_DetailWrapper>
         <p>{`오늘의 ${time}`}</p>
+
+        {/* 칼로리 디테일 */}
         <CalorieDetail />
+
         <S_FlexBox justifyContent="space-between" width="70%" height="50px">
           <p>총 칼로리</p>
           <p>{totalCalorie} kcal</p>
